pyproject.toml
[project]
name = "image-to-3dcad-sample"
version = "0.1.0"
description = "CAD Code Generation with Vision Language Models"
readme = "README.md"
requires-python = ">=3.13,<3.14"
dependencies = [
    # LLM/VLM dependencies
    "langchain>=1.2.14",
    "langchain-google-genai>=4.2.1",
    "langchain-google-vertexai>=3.2.2",
    "langgraph>=1.1.4",
    "google-genai>=1.69.0",
    "langchain-community>=0.4.1",
    "tiktoken>=0.12.0",
    "langchain-chroma>=1.1.0",
    "langchain-google-community>=3.0.5",
    # CAD dependencies
    "build123d>=0.10.0",
    "cadquery-ocp>=7.8.1",
    # Core dependencies
    "pillow>=12.1.1",
    "matplotlib>=3.10.8",
    "python-dotenv>=1.2.2",
    "pydantic>=2.12.5",
    "numpy>=2.4.4",
    "scipy>=1.17.1",
    # Google Cloud dependencies
    "google-cloud-discoveryengine>=0.18.0",
    "google-auth>=2.49.1",
    # Utilities
    "nest-asyncio>=1.6.0",
    "trimesh>=4.11.5",
    "pyrender>=0.1.45",
    "pyglet>=2.1.13",
    # PDF processing
    "pypdfium2>=5.6.0",
]

[dependency-groups]
dev = [
    "notebook",
    "ipython",
    "ruff~=0.12.12",
    "ty~=0.0.1a17",
    "pre-commit~=4.2.0",
    "pytest>=9.0.2",
]

[tool.uv]
package = true

[tool.hatch.build.targets.wheel]
packages = ["src/domain", "src/application", "src/infrastructure", "src/presentation"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.ruff]
line-length = 90
src = ["src"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long
    "B008",  # function calls in argument defaults
    "C901",  # too complex
    "W191",  # indentation contains tabs
    "B904",  # raise ... from ...
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
